LayoutTests/imported/w3c/web-platform-tests/workers/constructors/Worker/AbstractWorker.onerror-expected.txt

CONSOLE MESSAGE: Error: error from onerror.js

PASS AbstractWorker.onerror
